// src/entities/token.d.ts
type KuromojiToken = {
  allFeatures: string;
  allFeaturesArray: string[];
  baseForm: string;
  conjugationForm: string;
  conjugationType: string;
  known: boolean;
  partOfSpeechLevel1: string;
  partOfSpeechLevel2: string;
  partOfSpeechLevel3: string;
  partOfSpeechLevel4: string;
  position: number;
  pronunciation: string;
  reading: string;
  surface: string;
  user: boolean;
};

export default KuromojiToken